Add `createEnvironment.contentButton` setting (#21212)
Closes https://github.com/microsoft/vscode-python/issues/20982

---------
