+++ package.nls.json
@@ -25,6 +25,7 @@
     "python.command.python.analysis.restartLanguageServer.title": "Restart Language Server",
     "python.command.python.launchTensorBoard.title": "Launch TensorBoard",
     "python.command.python.refreshTensorBoard.title": "Refresh TensorBoard",
+    "python.createEnvironment.contentButton.description": "Show or hide Create Environment button in the editor for `requirements.txt` or other dependency files.",
     "python.menu.createNewFile.title": "Python File",
     "python.editor.context.submenu.runPython": "Run Python",
     "python.editor.context.submenu.runPythonInteractive": "Run in Interactive window",
