+++ src/client/extensionActivation.ts
@@ -54,7 +54,7 @@ import { DynamicPythonDebugConfigurationService } from './debugger/extension/con
 import { registerCreateEnvironmentFeatures } from './pythonEnvironments/creation/createEnvApi';
 import { IInterpreterQuickPick } from './interpreter/configuration/types';
 import { registerInstallFormatterPrompt } from './providers/prompts/installFormatterPrompt';
-import { registerPyProjectTomlCreateEnvFeatures } from './pythonEnvironments/creation/pyprojectTomlCreateEnv';
+import { registerCreateEnvButtonFeatures } from './pythonEnvironments/creation/createEnvButtonContext';
 
 export async function activateComponents(
     // `ext` is passed to any extra activation funcs.
@@ -98,7 +98,7 @@ export function activateFeatures(ext: ExtensionState, _components: Components): 
     );
     const pathUtils = ext.legacyIOC.serviceContainer.get<IPathUtils>(IPathUtils);
     registerCreateEnvironmentFeatures(ext.disposables, interpreterQuickPick, interpreterPathService, pathUtils);
-    registerPyProjectTomlCreateEnvFeatures(ext.disposables);
+    registerCreateEnvButtonFeatures(ext.disposables);
 }
 
 /// //////////////////////////
